Customize invoker user memory
Currently, all invokers share the same value from invoker.user-memory,
but in some case, some invoker machines have big memory, some invoker
machines may have low memory, so it is necessary to customize invoker
user-memory

+++ common/scala/src/main/scala/whisk/core/entity/Size.scala
@@ -109,7 +109,7 @@ object ByteSize {
   def fromString(sizeString: String): ByteSize = {
     val matcher = regex.matcher(sizeString)
     if (matcher.matches()) {
-      val size = matcher.group(1).toInt
+      val size = matcher.group(1).toLong
       val unit = matcher.group(2).charAt(0).toUpper match {
         case 'B' => SizeUnits.BYTE
         case 'K' => SizeUnits.KB
